Force GUI askpass to solve #6
git use 'get_terminal_prompt' to ask username/password when for example
user tried to push commits on a remote repository with https://

However, this 'get_terminal_prompt' does not work properly on neovim
and does not work on Vim 8 so an issue explained in #6 exist.

This commit force to use GUI based askpass program.
A GUI based askpass is required while the process may happens in
job.wait() loop.

Currently, Windows is not supported due to my knowledge limitation.
Hope someone make and send me a PR of askpass script which works well
on Windows.