Indicate with '--' that if the buffer is partial
When a command is called with -- {pathspec}..., the result is
partial (filtered with the pathspec).
So indicate that with buffer name otherwise users would mis-understand
that "Oh, there are only few files containing 'foo'" or so on.

Additionally, diff/show should use bufhidden=wipe when the result is
partial while same buffer name may contains a different result.