Replace Vim.Buffer.Anchor to more simple version
This commit change buffer opening behavior a bit.

Vim.Buffer.Anchor is broken in latest Vim and it's a bit complicated
to fix.
Additionally, the main reason for using Vim.Buffer.Anchor is to do
NOT open a buffer in small window such as VimFilerExplorer or
gina-status so using Vim.Buffer.Anchor is a bit overkill for that
purpose.

The new gina#core#locator is much simple which check the window size
to determine if the window is suitable to show a content and it seems
well enough for the purpose above.