Use exact match or suffix leading with =
Previous implementation returns wront index of '--color' when args contains values which
starts from '--color\W' (e.g. '--color-sidebar-text')

This commit change the pattern which allow

1. Exact match (e.g. '--color')
2. Exact match + suffix with leading = (e.g. '--color=always')

Additionally, when a pattern is given by a prompt, a leading '-e' is required when a
pattern starts from '-' so add '-e' always.

This commit fix #141